Refactor URLParamsHelper into TS service
